+++ Base/BaseComponent.py
@@ -1,0 +1,3 @@
+class Component:
+    def __init__(self, driver):
+        self.driver = driver
