+++ Search/SearchPage.py
@@ -1,0 +1,27 @@
+from Base.BasePage import Page
+from Common.CommonComponents import Tracks, Albums, Player
+from Search.SearchComponents import SearchBar, MainLayout
+
+
+class SearchPage(Page):
+    path = 'search'
+
+    @property
+    def search_bar(self):
+        return SearchBar(self.driver)
+
+    @property
+    def main_layout(self):
+        return MainLayout(self.driver)
+
+    @property
+    def tracks(self):
+        return Tracks(self.driver)
+
+    @property
+    def albums(self):
+        return Albums(self.driver)
+
+    @property
+    def player(self):
+        return Player(self.driver)
