+++ tests/login_test.py
@@ -2,57 +2,8 @@ import os
 import unittest
 
 from selenium.webdriver import DesiredCapabilities, Remote
-from selenium.webdriver.support.ui import WebDriverWait
-from selenium.webdriver.common.by import By
-from selenium.webdriver.support import expected_conditions as EC
 
-from tests import Page, Component
-
-
-class LoginPage(Page):
-    PATH = 'signin'
-
-    @property
-    def form(self):
-        return LoginForm(self.driver)
-
-
-class LoginForm(Component):
-    EMAIL = '//input[@name="email"]'
-    PASSWORD = '//input[@name="password"]'
-    LOGIN_BUTTON = '//input[@class="auth-form__submit"]'
-    FRONTEND_WARNINGS = '//div[@class="auth-form__invalidities form__invalidities"]'
-    BACKEND_WARNINGS_CLS = 'auth-form__fail_msg'
-
-    def set_email(self, email):
-        input = WebDriverWait(self.driver, 10, 0.1).until(
-            lambda d: d.find_element_by_xpath(self.EMAIL)
-        )
-        input.send_keys(email)
-
-    def set_password(self, password):
-        input = WebDriverWait(self.driver, 10, 0.1).until(
-            lambda d: d.find_element_by_xpath(self.PASSWORD)
-        )
-        input.send_keys(password)
-
-    def login(self):
-        button = WebDriverWait(self.driver, 10, 0.1).until(
-            lambda d: d.find_element_by_xpath(self.LOGIN_BUTTON)
-        )
-        button.click()
-
-    def frontend_warnings(self):
-        warnings = WebDriverWait(self.driver, 10, 0.1).until(
-            lambda d: d.find_element_by_xpath(self.FRONTEND_WARNINGS)
-        )
-        return warnings
-
-    def backend_warnings(self):
-        warnings = WebDriverWait(self.driver, 10, 0.1).until(
-            lambda d: d.find_element_by_class_name(self.BACKEND_WARNINGS_CLS)
-        )
-        return warnings
+from Login.LoginPage import LoginPage
 
 
 class LoginTest(unittest.TestCase):
@@ -82,10 +33,7 @@ class LoginTest(unittest.TestCase):
         assert not self.login_form.backend_warnings().text
 
         self.login_form.login()
-
-        WebDriverWait(self.driver, 10, 0.1).until(
-            EC.presence_of_element_located((By.CLASS_NAME, "avatar__img"))
-        )
+        self.login_form.check_login()
 
     def test_empty_form(self):
         self.login_form.login()
