+++ tests/register_test.py
@@ -3,71 +3,8 @@ import random
 import unittest
 
 from selenium.webdriver import DesiredCapabilities, Remote
-from selenium.webdriver.support.wait import WebDriverWait
-from selenium.webdriver.common.by import By
-from selenium.webdriver.support import expected_conditions as EC
 
-from tests import Page, Component
-
-
-class RegisterPage(Page):
-    PATH = 'signup'
-
-    @property
-    def form(self):
-        return RegisterForm(self.driver)
-
-
-class RegisterForm(Component):
-    NICKNAME = '//input[@name="nickname"]'
-    EMAIL = '//input[@name="email"]'
-    PASSWORD = '//input[@name="password"]'
-    CONFIRM_PASSWORD = '//input[@name="confirm_password"]'
-    REGISTER_BUTTON = '//input[@class="auth-form__submit"]'
-    FRONTEND_WARNINGS = '//div[@class="auth-form__invalidities"]'
-    BACKEND_WARNINGS_CLS = 'auth-form__fail_msg'
-
-    def set_nickname(self, nickname):
-        input = WebDriverWait(self.driver, 10, 0.1).until(
-            lambda d: d.find_element_by_xpath(self.NICKNAME)
-        )
-        input.send_keys(nickname)
-
-    def set_email(self, email):
-        input = WebDriverWait(self.driver, 10, 0.1).until(
-            lambda d: d.find_element_by_xpath(self.EMAIL)
-        )
-        input.send_keys(email)
-
-    def set_password(self, password):
-        input = WebDriverWait(self.driver, 10, 0.1).until(
-            lambda d: d.find_element_by_xpath(self.PASSWORD)
-        )
-        input.send_keys(password)
-
-    def set_confirm_password(self, password):
-        input = WebDriverWait(self.driver, 10, 0.1).until(
-            lambda d: d.find_element_by_xpath(self.CONFIRM_PASSWORD)
-        )
-        input.send_keys(password)
-
-    def register(self):
-        button = WebDriverWait(self.driver, 10, 0.1).until(
-            lambda d: d.find_element_by_xpath(self.REGISTER_BUTTON)
-        )
-        button.click()
-
-    def frontend_warnings(self):
-        warnings = WebDriverWait(self.driver, 10, 0.1).until(
-            lambda d: d.find_element_by_xpath(self.FRONTEND_WARNINGS)
-        )
-        return warnings
-
-    def backend_warnings(self):
-        warnings = WebDriverWait(self.driver, 10, 0.1).until(
-            lambda d: d.find_element_by_class_name(self.BACKEND_WARNINGS_CLS)
-        )
-        return warnings
+from Register.RegisterPage import RegisterPage
 
 
 class RegisterTest(unittest.TestCase):
@@ -100,10 +37,7 @@ class RegisterTest(unittest.TestCase):
         assert not self.register_form.backend_warnings().text
 
         self.register_form.register()
-
-        WebDriverWait(self.driver, 10, 0.1).until(
-            EC.presence_of_element_located((By.CLASS_NAME, "avatar__img"))
-        )
+        self.register_form.check_register()
 
     def test_empty_form(self):
         self.register_form.register()
